Removed IScheduler<TArg>. Removed Schedule<TArg> from interface, made extensions instead.
