+++ Vostok.Applications.Scheduled/IScheduler.cs
@@ -8,9 +8,4 @@ namespace Vostok.Applications.Scheduled
     {
         DateTimeOffset? ScheduleNext(DateTimeOffset from);
     }
-
-    [PublicAPI]
-    public interface IScheduler<TArg> : IScheduler
-    {
-    }
 }
