+++ Vostok.Applications.Scheduled/ScheduledActionsBuilder.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
-using Vostok.Applications.Scheduled.Schedulers;
 using Vostok.Logging.Abstractions;
 using Vostok.Tracing.Abstractions;
 
@@ -50,9 +49,6 @@ namespace Vostok.Applications.Scheduled
             return this;
         }
 
-        public IScheduledActionsBuilder Schedule<TArg>(string name, IScheduler<TArg> scheduler, Func<TArg, IScheduledActionContext, Task> payload, ScheduledActionOptions options) =>
-            Schedule(name, scheduler, WrapArgumentExtraction(payload, scheduler), options);
-
         internal ScheduledActionsRunner BuildRunnerInternal()
             => new ScheduledActionsRunner(actions.Select(action => new ScheduledActionRunner(action, log, tracer)).ToArray(), log);
 
@@ -63,19 +59,6 @@ namespace Vostok.Applications.Scheduled
                 action(context);
                 return Task.CompletedTask;
             };
-        }
-
-        private static TArg ExtractArgumentFromOnDemandScheduler<TArg>(IScheduler scheduler)
-        {
-            if (scheduler is OnDemandSchedulerWithArgument<TArg> onDemandScheduler)
-                return onDemandScheduler.GetLastArgumentValue();
-
-            throw new NotSupportedException("Argument passing is supported for OnDemandScheduler only.");
-        }
-
-        private static Func<IScheduledActionContext, Task> WrapArgumentExtraction<TArg>(Func<TArg, IScheduledActionContext, Task> action, IScheduler scheduler)
-        {
-            return context => action(ExtractArgumentFromOnDemandScheduler<TArg>(scheduler), context);
         }
     }
 }
