+++ Vostok.Applications.Scheduled/Schedulers/OnDemandScheduler.cs
@@ -63,29 +63,4 @@ namespace Vostok.Applications.Scheduled.Schedulers
 
         public override string ToString() => "OnDemand";
     }
-
-    internal class OnDemandSchedulerWithArgument<TArg> : OnDemandScheduler, IScheduler<TArg>
-    {
-        private TArg lastArgument;
-        
-        public void Demand(TArg argument)
-        {
-            lock (Sync)
-            {
-                lastArgument = argument;
-                base.Demand();
-            }
-        }
-
-        public Task DemandWithFeedback(TArg argument)
-        {
-            lock (Sync)
-            {
-                lastArgument = argument;
-                return base.DemandWithFeedback();
-            }
-        }
-
-        public TArg GetLastArgumentValue() => lastArgument;
-    }
 }
